CompareScreen: create a medium for writing comment about irregular readings in compare screen. Store them locally and send them to server in LinkedHashMap collection.
HttpService: create a method to make post request for storing comments
in LinkedHashMap collection to server.

CompareScreen: create a sub menu frame to choose tenant.

InputScreen: if meters are changed, reading value of old meters just before they are changed needs to be managed and stored as well. crate a medium to add those data to InputScreen and make other corresponding changes.