Client Side
Photos are now be able to attached.When attaching photos,separate copies
are created in app's directory.

When user jumps from IS01 to CS01,photos are also going to be sent to
server and stored in DB.

Created error display window and handled a few exception cases.

Created CheckMenu, yet to be completed.